+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.0.18] - 2018-10-24
 ### Added
 - `pf-expand-collapse` component.
 - `pf-vertical-nav-divider` component.
@@ -194,7 +194,8 @@ disabled, as per PatternFly design guidelines.
 - `pf-toolbar` component
 - `pf-utilization-bar-chart` component
 
-[Unreleased]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.17...HEAD
+[Unreleased]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.18...HEAD
+[0.0.18]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.17...v0.0.18
 [0.0.17]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.16...v0.0.17
 [0.0.16]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.15...v0.0.16
 [0.0.15]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.14...v0.0.15